requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
click==8.1.7
dnspython==2.6.1
environs==11.0.0
exceptiongroup==1.2.2
fastapi==0.112.0
fastapi-cache2==0.2.2
fastapi-pagination==0.12.26
h11==0.14.0
idna==3.7
marshmallow==3.21.3
packaging==24.1
pendulum==3.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
time-machine==2.14.2
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.5
